tune: label: fix bitmap entry corruption when adding new volume label
When adding new label using tune.exfat or exfatlabel against exfat
device formatted by some camera vendor, bitmap entry in root entry is
corrupted by overwriting. the format utils of vendor doesn't add volume
entry. tune and exfatlabel of exfatprogs assumes that the first entry of
the root entry is a volume entry and try overwrite it. This patch lookup
and updates the volume entry in the root entry. And adds a new entry to
an empty slot if it doesn't exist.

Signed-off-by: Namjae Jeon <linkinjeon@kernel.org>